change: add serializable UId to user and profile
