+++ src/main/java/me/universi/image/exceptions/ImageException.java
@@ -1,0 +1,8 @@
+package me.universi.image.exceptions;
+
+public class ImageException extends Exception {
+
+    public ImageException(String m) {
+        super(m);
+    }
+}
