+++ src/main/java/me/universi/user/entities/User.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
+import java.io.Serial;
+import java.io.Serializable;
 import me.universi.profile.entities.Profile;
 import me.universi.user.enums.Authority;
 import me.universi.user.services.JsonEmailOwnerSessionFilter;
@@ -17,7 +19,10 @@ import java.util.Collection;
 import java.util.UUID;
 
 @Entity(name = "system_users")
-public class User implements UserDetails {
+public class User implements UserDetails, Serializable {
+
+    @Serial
+    private static final long serialVersionUID = -4463545345342344343L;
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
